+++ CHANGELOG.md
@@ -47,7 +47,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Avoid using the deprecated `GeneralUtility::_GP` (#2827)
 - Replace deprecated `setUpBackendUserFromFixture` calls in tests (#2820)
 - Avoid the deprecated `cruser_id` (#2821)
-- Avoid using the deprecated `ObjectManager` (#2819, #2871, #2972)
+- Avoid using the deprecated `ObjectManager` (#2819, #2871, #2972, #2976)
 - Stop using the deprecated `Mock::setMethods()` (#2815)
 - Avoid usage of the deprecated `strftime` (#2814)
 - Avoid warnings in the tests with PHP 8.3 (#2812)